Conflicts:
	src/gromacs/CMakeLists.txt
	src/gromacs/gmxana/addconf.c
	src/gromacs/gmxana/calcpot.c
	src/gromacs/gmxana/gmx_disre.c
	src/gromacs/gmxlib/atomprop.c
	src/gromacs/gmxpreprocess/convparm.c
	src/gromacs/gmxpreprocess/pdb2top.cpp
	src/gromacs/legacyheaders/convparm.h
	src/gromacs/legacyheaders/force.h
	src/gromacs/legacyheaders/pdb2top.h
	src/gromacs/mdlib/forcerec.c
	src/ngmx/g_showcol.c
	src/programs/gmx/convparm.c
	src/programs/gmx/convparm.h
	src/programs/gmx/g_x2top.c
	src/programs/gmx/pdb2gmx.c
	src/programs/grompp/convparm.c
	src/programs/grompp/convparm.h
	src/programs/mdrun/runner.c

+++ src/gromacs/legacyheaders/macros.h
@@ -66,6 +66,6 @@
 #endif
 
 /* This macro calculates the size of a array */
-#define asize(a) (sizeof(a)/sizeof((a)[0]))
+#define asize(a) ((int)(sizeof(a)/sizeof((a)[0])))
 
 #endif  /* _macros_h */
